+++ lib/gilded_rose.rb
@@ -2,30 +2,56 @@ class GildedRose
 	SULFURAS = "Sulfuras, Hand of Ragnaros"
 	BRIE = "Aged Brie"
 	CONCERT_PASS = "Backstage passes to a TAFKAL80ETC concert"
-	UPPER_QLTY_LIMIT = 50
-	LOWER_QLTY_LIMIT = 0
-	CONCERT_QUALITY_DOUBLE_LIMIT = 10
-	CONCERT_QUALITY_TRIPLE_LIMIT = 5
-	CONCERT_EXPIRY_VALUE = 0
+	MAX_QLTY_LIMIT = 50
+	MIN_QLTY_LIMIT = 0
+	DAYS_LEFT_TO_DOUBLE_CONCERT_QLTY = 10
+	DAYS_LEFT_TO_TRIPLE_CONCERT_QLTY = 5
+	CONCERT_EXPIRY_QLTY = 0
+	DOUBLE = 2
+	TRIPLE = 3
+	SINGLE = 1
 
 	def initialize(items)
 		@items = items
 	end
 
-=begin
-This function takes an item and a qualityLimit, and increments the quality of the item if it hasn't reached the limit 
-=end
+
+	#This function takes an item and a qualityLimit, and increments the quality of the item if it hasn't reached the limit 
 	def incrementQuality(item, qualityLimit)
 		if item.quality < qualityLimit
 			item.quality += 1
 		end
 	end 
 
+	# def incrementDoubleQuality(item, qualityLimit)
+	# 	incrementQuality(item, qualityLimit)
+	# 	incrementQuality(item, qualityLimit)
+	# end
+
+	# increment quality n times
+	def incrementQualityMultiple(item, qualityLimit, n)
+		i = 0
+		while i < n
+			incrementQuality(item, qualityLimit)
+			i += 1
+		end
+	end
+
 	def decrementQuality(item, qualityLimit)
 		if item.quality > qualityLimit
 			item.quality -= 1
 		end
 	end
+
+	# decrement Quality multiple times
+	def decrementQualityMultiple(item, qualityLimit, n)
+		i = 0
+		while i < n
+			decrementQuality(item, qualityLimit)
+			i += 1
+		end
+	end
+
 	# update the sell in value of a particular item
 	def decrementSellIn(item)
 		item.sell_in -=  1
@@ -36,39 +62,54 @@ This function takes an item and a qualityLimit, and increments the quality of th
 		item.sell_in < 0
 	end
 
+	def updateBrie(item)
+		# if Brie has expired, increment its quality
+		if hasExpired(item)
+			incrementQualityMultiple(item, MAX_QLTY_LIMIT, DOUBLE)
+		else
+			incrementQuality(item, MAX_QLTY_LIMIT)
+		end
+	end
+	
+	def updateConcertPasses(item)
+		# if there are less thatn DAYS_LEFT_TO_TRIPLE_CONCERT_QLTY days to concert, decrement again
+		if item.sell_in < DAYS_LEFT_TO_TRIPLE_CONCERT_QLTY
+			incrementQualityMultiple(item, MAX_QLTY_LIMIT, TRIPLE)
+		# if there are less than DAYS_LEFT_TO_DOUBLE_CONCERT_QLTY days to concert, decrement again
+		elsif item.sell_in < DAYS_LEFT_TO_DOUBLE_CONCERT_QLTY
+			incrementQualityMultiple(item, MAX_QLTY_LIMIT, DOUBLE)
+		# increase normally
+		else
+			incrementQuality(item, MAX_QLTY_LIMIT)
+		end
+		
+		# if concert passes has expired, its quality is 0
+		if hasExpired(item)
+			item.quality = CONCERT_EXPIRY_QLTY
+		end
+	end
+
+	def updateNormalItem(item)
+		# if a normal item has expired, quality degrades twice as fast
+		if hasExpired(item)
+			decrementQualityMultiple(item, MIN_QLTY_LIMIT, DOUBLE)
+		else
+			decrementQuality(item,MIN_QLTY_LIMIT)
+		end
+	end
+
 	def update_quality
 		@items.each do |item|
 			if item.name == SULFURAS
 				next
 			end
-
 			decrementSellIn(item)
 			if item.name == BRIE
-				incrementQuality(item, UPPER_QLTY_LIMIT)
-				# if Brie has expired, increment its quality
-				if hasExpired(item)
-					incrementQuality(item, UPPER_QLTY_LIMIT)
-				end
+				updateBrie(item)
 			elsif item.name == CONCERT_PASS
-				incrementQuality(item, UPPER_QLTY_LIMIT)
-				# if there are less than CONCERT_QUALITY_DOUBLE_LIMIT days to concert, decrement again
-				if item.sell_in < CONCERT_QUALITY_DOUBLE_LIMIT
-					incrementQuality(item, UPPER_QLTY_LIMIT)
-				end
-				# if there are less thatn CONCERT_QUALITY_TRIPLE_LIMIT days to concert, decrement again
-				if item.sell_in < CONCERT_QUALITY_TRIPLE_LIMIT
-					incrementQuality(item, UPPER_QLTY_LIMIT)
-				end
-				# if concert passes has expired, its quality is 0
-				if hasExpired(item)
-					item.quality = CONCERT_EXPIRY_VALUE
-				end
+				updateConcertPasses(item)
 			else
-				decrementQuality(item,LOWER_QLTY_LIMIT)
-				# if it is a normal item, decrement twice as fast
-				if hasExpired(item)
-					decrementQuality(item,LOWER_QLTY_LIMIT)
-				end
+				updateNormalItem(item)
 			end
 		end
 	end